Move iofuncs and miscfuncs to root
