+++ analyze/checkerflickerplus.py
@@ -13,9 +13,9 @@ import os
 import warnings
 
 from .. import randpy
+from .. import misc as msc
 from ..modules import analysisfuncs as asc
-from ..modules import iofuncs as iof
-from ..modules import miscfuncs as msc
+from .. import io as iof
 from ..plot import util as plf
