+++ analyze/omb.py
@@ -12,7 +12,7 @@ from scipy.stats.mstats import mquantiles
 
 from .. import randpy
 from ..modules import analysisfuncs as asc
-from ..modules import iofuncs as iof
+from .. import io as iof
 from ..plot import util as plf
