+++ modules/__init__.py
@@ -1,8 +1,5 @@
 from . import analysis_scripts as analysisfuncs
-from . import iofuncs, miscfuncs
 
 __all__ = [
     'analysisfuncs',
-    'iofuncs',
-    'miscfuncs',
 ]
