+++ plot/checkerstas.py
@@ -11,7 +11,7 @@ import os
 
 from . import util as plf
 from ..modules import analysisfuncs as asc
-from ..modules import iofuncs as iof
+from .. import io as iof
 
 
 def checkerstas(exp_name, stim_nr, filename=None):
